Added return values and fixed main function code
Forgot to put return values for all of the corresponding values of the band colors. Also, the main function is calling any of the three functions (three_band, four_band, and five_band) twice, because I placed the function calls in such a way that the if statements triggered the functions to test if the functions returned null values for incorrect inputs. And then the else statement fired, which should display the results from the functions, but I forgot to put print() functions.